back-end/requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.0
django-cors-headers==4.4.0
django-jazzmin==3.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysql-connector-python==9.0.0
mysqlclient==2.2.4
pillow==10.4.0
PyJWT==2.9.0
reportlab==4.2.2
sqlparse==0.5.1
tzdata==2024.1